fix(cli): init templates still assume the same versions (#33449)
The init templates still have a single `%cdk-version%` placeholder for both the library and the CLI.

Introduce a new `%cdk-cli-version%` placeholder and fill it with the "current" version of the CLI used to create the template.

Turning `DISPLAY_VERSION` into a lazy function, because we added a dependency `init.ts -> version.ts`, and since `init.ts` is included in a pre-build script, we now include `version.ts` in that pre-build script which tries to read a file at load time that still needs to be generated.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*