chore: remove service spec yarn resolutions (#33485)
These resolutions prevent us from fetching the latest schema. They were [added](https://github.com/aws/aws-cdk/pull/33481) because there are multiple breaking changes in the spec we weren't ready to accept yet. 

### Are we ready to accept the new spec?

Not quite.

### So whats the plan?

We need to remove these resolutions in order to get a service spec update PR that at least **surfaces** all the breaking changes in order to work on them. So the plan is:

1. Merge this PR.
2. Run the spec update workflow
3. Slap a `pr/do-not-merge` label on the resulting PR
4. Figure out what to do with the breaking changes it show.
5. Eventually merge the spec update after applying patches or allowing breaking changes to go through.

From a local execution at the time of writing this PR, those breaking changes should be:

```console
├[~] service aws-cognito
│ └ resources
│    └[~]  resource AWS::Cognito::UserPoolDomain
│       └ attributes
│          └[-] Id: string


├[~] service aws-dynamodb
│ └ resources
│    └[~]  resource AWS::DynamoDB::GlobalTable
│       └ properties
│          └[-] PointInTimeRecoverySpecification: PointInTimeRecoverySpecification


│    ├[~]  resource AWS::ECS::Service
│    │  ├ properties
│    │  │  ├ AvailabilityZoneRebalancing: (documentation changed)
│    │  │  └ CapacityProviderStrategy: (documentation changed)
│    │  └ types
│    │     └[~] type ServiceManagedEBSVolumeConfiguration
│    │       └ properties
│    │          └[-] VolumeInitializationRate: integer


│    └[~]  resource AWS::RDS::GlobalCluster
│       ├ properties
│       │  └[-] GlobalEndpoint: GlobalEndpoint
│       └ attributes
│          └[+] GlobalEndpoint: GlobalEndpoint

```

### Wait but you just detailed the breaking changes, can't we just work based of that?

No, we shouldn't be operating on the basis of local executions. The local execution diff is just to get a sense and show that there are indeed a few breaking changes already present. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*