chore: tweak enum-like parser to exclude certain classes (#33536)
### Issue # (if applicable)

N/A

### Reason for this change

The enum-like parser was pulling in some classes that did not follow the pattern of enum-like classes. This alleviates that by, rather than looking for fields that are strictly not strings, fields that strictly follow the pattern:
```ts
  public static readonly NEW_EXPRESSION = new SomeClass(...);
  public static readonly CALL_EXPRESSION = SomeClass.method(...);
  public static readonly ACCESS_EXPRESSION = MyClass.ANOTHER_ENUM;
```

### Description of changes

Altered the parsing logic to be more specific, and reran the script to parse enum-likes.

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

`yarn test`

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*