fix(eks-v2-alpha): can't delete fargate cluster (#33573)
### Issue # (if applicable)

Closes #33347.

### Reason for this change

The fargate cluster deletion issue is because the cluster admin access entry is deleted before deleting `KubernetesPatch`. 

Since deleting `KubernetesPatch` requires applying a restore patch to the cluster, it will still need the cluster access. In this case, because the access entry is deleted, kubectl provider won't be able to apply the patch to the cluster anymore.

### Description of changes

add an explicit dependency from patch to the access entry so the patch will only be deleted after the access entry

### Description of how you validated changes

unit tests/integration tests

### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*