fix(core): pressing Ctrl-C when content is bundled leaves broken asset (#33692)
When a bundling command is interrupted with Ctrl-C, the asset output directory has already been created. On the next synthesis, we assume the asset has already successfully been produced, don't do any bundling, and upload it.

We will then have produced and uploaded a broken asset.

Instead, the common pattern to handle this is:

- Do the work into a temporary directory
- Rename the temporary directory to the target directory only if the work succeeds.

Closes #33201, closes #32869, relates to #14474.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*