feat(rds): allow to specify availability zone for Aurora instances (#33515)
### Issue # (if applicable)

Closes #33503.
Closes #30618.

### Reason for this change

When creating a single-zone Aurora cluster and an EC2 instance in development purpose, I want to *pin* the Aurora instance and the EC2 instance in a same availability zone.
We can specify `AvailabilityZone` in Aurora's `AWS::RDS::DBInstance` as same as standalone RDS instances.

### Description of changes

Added `availabilityZone` prop in `ClusterInstanceOptions`

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

Unit test and integ test.
The integ test also verifies created instances are placed in expected availability zone.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*