feat(cx-api): declare support for CDK_TOOLKIT_VERSION env var (#33963)
### Reason for this change

CDK apps have access to an informative env variable, containing the version of the CLI running the app. With the release of the CDK Toolkit Library, this value does not always represent what it should. As a future replacement, we are introducing a new env variable that will contain the version and package name of the package

### Description of changes

Add a new env var name to `cx-api` to support a generic toolkit version. This value will contain the package name and version:

"<package>@x.y.z", e.g  "aws-cdk@2.1006.0" or "@aws-cdk/toolkit-lib@1.0.0"

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

n/a

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*